Fail if creation or deletion of a network on a device fails
When creating or deleting a network, NGS will add or remove that network
on all configured devices. Currently, a failure on any of those devices
will not cause the operation to fail. This can leave the network in an
inconsistent state.

This change modifies this behaviour, such that a failure on any device
will cause the operation to fail. If creation of a network fails,
neutron ML2 plugin will call the driver's delete methods to clean up.
Failed deletions are not cleaned up by the ML2 plugin, but this probably
makes more sense.

Change-Id: I9a26890df0b430a8181abd86c4a6a64d95eb4121
Story: 2006222
Task: 35820